Home Service working

+++ lib/src/app/modules/home/domain/entity/home_location_entity.dart
@@ -3,7 +3,7 @@ class HomeLocationEntity {
   String? name;
   String? address;
   String? imagePath;
-  int? averageGrade;
+  double? averageGrade;
   int? reviewsQnt;
 
   HomeLocationEntity({
