+++ lib/src/service/api/modules/home/response/response_get_best_rated_locations.dart
@@ -56,7 +56,7 @@ class ResponseGetBestRatedLocationsLocation {
   String? name;
   String? endereco;
   String? imgUrl;
-  int? averageGrade;
+  double? averageGrade;
   int? reviewsQnt;
 
   ResponseGetBestRatedLocationsLocation({
